+++ src/api/routes.py
@@ -4,6 +4,7 @@ This module takes care of starting the API Server, Loading the DB and Adding the
 from flask import Flask, request, jsonify, url_for, Blueprint
 from api.models import db, User
 from forms import UserForm
+from flask_jwt_extended import create_access_token
 
 api = Blueprint('api', __name__)
 
@@ -28,3 +29,21 @@ def create_user():
     else:
         errors = {field: errors[0] for field, errors in form.errors.items()}
         return jsonify({'error': 'validation error', 'errors': errors}), 400
+
+@api.route('/login', methods=['POST'])
+def handle_login():
+    email = request.json.get("email", None)
+    password = request.json.get("password", None)
+    if email != "test" or password != "test":
+        return jsonify({"msg": "Bad email or password"}), 401
+
+    access_token = create_access_token(identity=email)
+    return jsonify(access_token=access_token)
+
+@api.route('/user/account', methods=['POST'])
+def handle_account():
+        return 'Succesfully log in'
+
+@api.route('/logout')
+def logout():
+        return 'You log out'
